refactor: prevent layout components from unmounting
