+++ site/gatsby-browser.js
@@ -3,3 +3,4 @@ import "prismjs/themes/prism.css"
 import "./src/styles/global.scss"
 import "./src/styles/prism.scss"
 import "./src/styles/docs.scss"
+export { default as wrapPageElement } from "./gatsby/wrapPageElement"
