+++ site/gatsby/wrapPageElement.js
@@ -1,0 +1,9 @@
+// eslint-disable-next-line no-use-before-define
+import * as React from "react"
+import Layout from "../src/components/layout"
+
+const wrapPageElement = ({ element, props }) => {
+  return <Layout {...props}>{element}</Layout>
+}
+
+export default wrapPageElement
